refactor: changed project structure
